Refactor: find-tasks tabs and filters components
